pyproject.toml
[project]
name = "aidevs4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.1",
    "geopy>=2.4.1",
    "httpx>=0.28.1",
    "langchain>=1.2",
    "langchain-openai>=1.1",
    "langfuse>=4.0.0",
    "mcp>=1.26.0",
    "ngrok>=1.7.0",
    "pydantic-ai>=1.68.0",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "rich>=14.3.3",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "ruff>=0.15",
]

[tool.ruff]
target-version = "py314"
line-length = 120

[tool.ruff.lint]
select = [
    "E", "F", "W",       # pycodestyle + pyflakes
    "I",                  # isort
    "UP",                 # pyupgrade
    "B",                  # bugbear
    "SIM",                # simplify
    "RUF",                # ruff-specific
    "N",                  # pep8-naming
    "ANN",                # annotations
    "S",                  # bandit (security)
    "C4",                 # comprehensions
    "PIE",                # misc lint
    "PERF",               # performance
    "T20",                # print() detection
    "ISC",                # implicit string concat
    "DTZ",                # timezone-aware datetime
    "RET",                # return patterns
    "PTH",                # pathlib over os.path
    "PT",                 # pytest style
    "ARG",                # unused arguments
    "EM",                 # exception messages
    "TRY",                # exception handling
    "TCH",                # TYPE_CHECKING imports
    "FA",                 # future annotations
]
ignore = [
    "ANN401",  # Any type allowed sometimes
    "ISC001",  # conflicts with ruff formatter
    "TRY003",  # long exception messages OK
    "EM101",   # string literals in exceptions OK
    "RET504",  # unnecessary assignment before return
    "S101",    # assert is fine (we're not a library)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "ARG", "ANN"]
"settings.py" = ["S105"]  # default values for local dev are not real secrets

[tool.ruff.lint.isort]
combine-as-imports = true
known-first-party = ["aidevs4"]

[tool.pyright]
typeCheckingMode = "strict"
stubPath = "typings"
reportUnknownMemberType = "error"
reportUnknownArgumentType = "error"
reportUnknownVariableType = "error"
reportUnknownParameterType = "error"
reportUnknownLambdaType = "error"
reportMissingTypeStubs = "error"
